refactor(dimensions): move dimensions/margin types to /utils
