+++ src/lib/axes/axis_utils.ts
@@ -13,7 +13,7 @@ import {
   AxisStyle,
 } from '../series/specs';
 import { AxisConfig, Theme } from '../themes/theme';
-import { Dimensions, Margins } from '../utils/dimensions';
+import { Dimensions, Margins } from '../../utils/dimensions';
 import { AxisId, GroupId } from '../../utils/ids';
 import { Scale } from '../../utils/scales/scales';
 import { BBox, BBoxCalculator } from '../../utils/bbox/bbox_calculator';
