+++ src/state/annotation_utils.test.ts
@@ -9,7 +9,7 @@ import {
   Rotation,
 } from '../lib/series/specs';
 import { DEFAULT_ANNOTATION_LINE_STYLE } from '../lib/themes/theme';
-import { Dimensions } from '../lib/utils/dimensions';
+import { Dimensions } from '../utils/dimensions';
 import { AnnotationId, AxisId, getAnnotationId, getAxisId, getGroupId, GroupId } from '../utils/ids';
 import { ScaleBand } from '../utils/scales/scale_band';
 import { ScaleContinuous } from '../utils/scales/scale_continuous';
