+++ src/state/annotation_utils.ts
@@ -17,7 +17,7 @@ import {
   Rotation,
 } from '../lib/series/specs';
 import { LineAnnotationStyle } from '../lib/themes/theme';
-import { Dimensions } from '../lib/utils/dimensions';
+import { Dimensions } from '../utils/dimensions';
 import { AnnotationId, AxisId, GroupId } from '../utils/ids';
 import { Scale, ScaleType } from '../utils/scales/scales';
 import { Point } from './chart_state';
