+++ src/state/crosshair_utils.ts
@@ -1,5 +1,5 @@
 import { Rotation } from '../lib/series/specs';
-import { Dimensions } from '../lib/utils/dimensions';
+import { Dimensions } from '../utils/dimensions';
 import { Scale } from '../utils/scales/scales';
 import { isHorizontalRotation } from './utils';
