+++ src/utils/dimensions.ts
@@ -1,0 +1,13 @@
+export interface Dimensions {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+}
+
+export interface Margins {
+  top: number;
+  bottom: number;
+  left: number;
+  right: number;
+}
